Introduce options and options builder
Introduced a builder pattern to keep the code easy to read by
encapsulating Dependency Check execution arguments and options for the
code execution at all in dedicated objects.
So a singel point of responsibility was created and helps t o keep track
of changes and makes it easier to extend the supported
arguments/options.