Update/Download Dependency Check data
Extended the task to execte the DC to update the data only. No scan is
executed. In addition it's possible to define if and where the DC
execution logfile should be stored.
Furthermore, the default path was adjusted to store all files in the
Azure DevOps artifacts folder "/a" instead of in the "/TestResults"
folder.